requirements.txt
streamlit
pandas
requests
Pillow
easyocr
numpy
pyzbar
opencv-python-headless
openpyxl